minor code refactor
